[feat] 카드목록 보기 별도로 분리
